DeepBlocker/requirements.txt
# For DeepBlocker
fasttext-wheel
numpy
pandas
pathlib
scipy
scikit-learn
torch
torchtext
